requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.5.0
    # via
    #   starlette
    #   watchgod
asgiref==3.5.1
    # via uvicorn
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via uvicorn
fastapi==0.76.0
    # via -r requirements.in
google-api-core==2.7.3
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.6.6
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.0
    # via google-cloud-storage
google-cloud-storage==2.3.0
    # via -r requirements.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-storage
googleapis-common-protos==1.56.0
    # via google-api-core
h11==0.13.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
protobuf==3.20.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.9.0
    # via fastapi
python-dotenv==0.20.0
    # via uvicorn
pyyaml==6.0
    # via uvicorn
requests==2.27.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
rsa==4.8
    # via google-auth
six==1.16.0
    # via google-auth
sniffio==1.2.0
    # via anyio
starlette==0.18.0
    # via fastapi
typing-extensions==4.2.0
    # via pydantic
urllib3==1.26.9
    # via requests
uvicorn[standard]==0.17.6
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.8.2
    # via uvicorn
websockets==10.3
    # via uvicorn
